Create PlayerRating during omniauth login

+++ db/migrate/20231201053808_create_historical_player_ratings.rb
@@ -4,8 +4,8 @@ class CreateHistoricalPlayerRatings < ActiveRecord::Migration[6.1]
       t.string :player_name, comment: "historical player name"
       t.references :player, index: true, null: true
       t.integer :elo, comment: "trueskill mu normalized between 1000 and 2000"
-      t.decimal :mu, comment: "trueskill mu"
-      t.decimal :sigma, comment: "trueskill sigma"
+      t.float :mu, comment: "trueskill mu"
+      t.float :sigma, comment: "trueskill sigma"
       t.date :last_played, comment: "last played match"
 
       t.timestamps
